+++ frontend/project/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for pepsi project.
+Django settings for project.
 
 Generated by 'django-admin startproject' using Django 5.2.3.
