+++ frontend/project/urls.py
@@ -1,5 +1,5 @@
 """
-URL configuration for pepsi project.
+URL configuration for project.
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/5.2/topics/http/urls/
