+++ frontend/static/templates/navbar.html
@@ -5,7 +5,7 @@
       <b>PEPSI Dashboard</b>
     </a>
     <div class="ms-auto">
-      <a class="nav-link text-white p-0" href="https://github.com/ronjakrg/thesis-pepsi-package/">
+      <a class="nav-link text-white p-0" href="https://github.com/ronjakrg/pepsipy/">
         <img src="{% static 'img/github.svg' %}" alt="GitHub" width="30" height="30" class="d-inline-block align-text-top">
       </a>
     </div>
