+++ src/pepsipy/__main__.py
@@ -3,9 +3,9 @@
 #   This file is just an experimental sandbox and is
 #   not part of the implementation.
 #
-#   To run this code, execute python -m pepsi.
+#   To run this code, execute python -m pepsipy.
 # ──────────────────────────────────────────────────────
 
-from pepsi import Calculator
+from pepsipy import Calculator
 
 Calculator.hello_pepsi()
