+++ src/pepsipy/external/ipc-2.0.1/scripts/ipc2_lib/essentials.py
